+++ R/zzz.R
@@ -27,6 +27,9 @@
     # Version of 'extendr_api' to be used
     rextendr.extendr_deps = list(
       `extendr-api` = "*"
+    ),
+    rextendr.extendr_dev_deps = list(
+      `extendr-api` = list(git = "https://github.com/extendr/extendr")
     )
   )